agents/requirements.txt
# Core dependencies first (install in order to avoid backtracking)
pydantic==2.9.0
pyyaml==6.0.2
typing-extensions==4.12.2

# LangChain ecosystem - install core first
# langchain-anthropic 0.3.15 requires langchain-core>=0.3.63
# langchain 0.3.15 requires langchain-core>=0.3.31,<0.4.0
# langgraph 0.2.45 requires langchain-core>=0.2.43,<0.4.0
# langchain-openai 0.2.8 requires langchain-core>=0.3.17,<0.4.0
# Use minimum version that satisfies all: 0.3.63
langchain-core>=0.3.63,<0.4.0
langchain==0.3.15
langgraph==0.2.45

# LangChain integrations
langchain-openai==0.2.8
langchain-anthropic==0.3.15
langchain-google-genai==2.0.7

# LLM provider SDKs
openai==1.54.0
# langchain-anthropic 0.3.15 requires anthropic>=0.52.0
anthropic>=0.52.0,<1.0.0
google-generativeai==0.8.3

# WebSocket client
websockets==13.1

# ML dependencies (lightweight)
scikit-learn==1.3.2
numpy==1.24.3

